+++ src/Pim/Bundle/DataGridBundle/Extension/Filter/OrmFilterExtension.php
@@ -2,8 +2,6 @@
 
 namespace Pim\Bundle\DataGridBundle\Extension\Filter;
 
-use Oro\Bundle\FilterBundle\Grid\Extension\OrmFilterExtension as OroOrmFilterExtension;
-use Oro\Bundle\FilterBundle\Grid\Extension\Configuration;
 use Oro\Bundle\DataGridBundle\Datagrid\Common\DatagridConfiguration;
 use Oro\Bundle\DataGridBundle\Datagrid\Builder;
 use Pim\Bundle\DataGridBundle\Datasource\Orm\OrmDatasource as PimOrmDatasource;
@@ -15,19 +13,15 @@ use Pim\Bundle\DataGridBundle\Datasource\Orm\OrmDatasource as PimOrmDatasource;
  * @copyright 2013 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class OrmFilterExtension extends OroOrmFilterExtension
+class OrmFilterExtension extends AbstractFilterExtension
 {
     /**
-     * {@inheritdoc}
+     * @param DatagridConfiguration $config
+     *
+     * @return boolean
      */
-    public function isApplicable(DatagridConfiguration $config)
+    protected function matchDatasource(DatagridConfiguration $config)
     {
-        $filters = $config->offsetGetByPath(Configuration::COLUMNS_PATH, []);
-
-        if (!$filters) {
-            return false;
-        }
-
         $datasourceType = $config->offsetGetByPath(Builder::DATASOURCE_TYPE_PATH);
 
         return ($datasourceType == PimOrmDatasource::TYPE);