+++ CHANGELOG.md
@@ -13,7 +13,7 @@
 
 ## BC breaks
 - Add an argument HydratorInterface in ProductDatasource constructor (MongoDBODM support)
-- Add an argument $adapterClass (string for FQCN) in ProductFilterExtension constructor (MongoDBODM support)
+- Add an argument $adapterClass (string for FQCN) in ProductFilterExtension and OrmFilterExtension constructors (MongoDBODM support)
 - Remove deprecated fallback property of Locale entity
 
 # 1.0.0 - "Hare We Go" (2014-03-06)